+++ docbook/src/main/scala/org/opentorah/docbook/plugin/ProcessDocBookTask.scala
@@ -18,7 +18,10 @@ import scala.beans.BeanProperty
 import scala.collection.JavaConverters._
 import scala.xml.{Comment, Elem}
 
+// Note: Task class can not be final for Gradle to be able to decorate it.
 class ProcessDocBookTask extends DefaultTask {
+  setDescription(s"Process DocBook")
+  setGroup("publishing")
 
   private val layout: Layout = Layout.forProject(getProject)
   private val logger: Logger = getProject.getLogger